LayoutTests/crypto/subtle/aes-cbc-invalid-length-expected.txt

Test an unsupported AES key length.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a raw AES key from string literal...
PASS key.type is 'secret'
PASS key.extractable is true
PASS key.algorithm.name is 'aes-cbc'
PASS key.algorithm.length is 176
PASS key.usages is ['encrypt', 'decrypt']
Using the key to encrypt plaintext...
Failed, as expected. Note that the spec doesn't appear to clearly define which step should fail.
PASS successfullyParsed is true

TEST COMPLETE
